Fixing keyword argument
--HG--
branch : yt-3.0

+++ yt/data_objects/selection_data_containers.py
@@ -363,7 +363,7 @@ class YTCuttingPlaneBase(YTSelectionContainer2D):
             x = self._current_chunk.fcoords[:,0] - self.center[0]
             y = self._current_chunk.fcoords[:,1] - self.center[1]
             z = self._current_chunk.fcoords[:,2] - self.center[2]
-            tr = YTArray(x.size, dtype='float64', "code_length",
+            tr = YTArray(x.size, "code_length",
                          registry = self.pf.unit_registry)
             tr += x * self._x_vec[0]
             tr += y * self._x_vec[1]
@@ -373,7 +373,7 @@ class YTCuttingPlaneBase(YTSelectionContainer2D):
             x = self._current_chunk.fcoords[:,0] - self.center[0]
             y = self._current_chunk.fcoords[:,1] - self.center[1]
             z = self._current_chunk.fcoords[:,2] - self.center[2]
-            tr = YTArray(x.size, dtype='float64', "code_length",
+            tr = YTArray(x.size, "code_length",
                          registry = self.pf.unit_registry)
             tr += x * self._y_vec[0]
             tr += y * self._y_vec[1]
@@ -383,7 +383,7 @@ class YTCuttingPlaneBase(YTSelectionContainer2D):
             x = self._current_chunk.fcoords[:,0] - self.center[0]
             y = self._current_chunk.fcoords[:,1] - self.center[1]
             z = self._current_chunk.fcoords[:,2] - self.center[2]
-            tr = YTArray(x.size, dtype='float64', "code_length",
+            tr = YTArray(x.size, "code_length",
                          registry = self.pf.unit_registry)
             tr += x * self._norm_vec[0]
             tr += y * self._norm_vec[1]
